SLA-8762: renames image-uploader to image-search-ai-image-uploader
