+++ src/SprykerEco/Yves/ImageSearchAi/Theme/default/components/molecules/image-search-ai-image-uploader/image-search-ai-image-uploader.twig
@@ -1,8 +1,8 @@
 {% extends model('component') %}
 
 {% define config = {
-    name: 'image-uploader',
-    tag: 'image-uploader',
+    name: 'image-search-ai-image-uploader',
+    tag: 'image-search-ai-image-uploader',
 } %}
 
 {% define attributes = {
