+++ src/SprykerEco/Yves/ImageSearchAi/Theme/default/components/molecules/image-search-ai-image-uploader/index.ts
@@ -1,0 +1,12 @@
+import register from 'ShopUi/app/registry';
+import './image-search-ai-image-uploader.scss';
+
+export default register(
+    'eco-image-uploader',
+    () =>
+        import(
+            /* webpackMode: "lazy" */
+            /* webpackChunkName: "image-search-ai-image-uploader" */
+            './image-search-ai-image-uploader'
+            ),
+);
