+++ src/SprykerEco/Yves/ImageSearchAi/Theme/default/components/molecules/image-uploader-popup/image-uploader-popup.twig
@@ -11,7 +11,7 @@
 } %}
 
 {% block content %}
-    {% include molecule('image-uploader', 'ImageSearchAi') only %}
+    {% include molecule('image-search-ai-image-uploader', 'ImageSearchAi') only %}
 {% endblock %}
 
 {% block body %}
