+++ src/SprykerEco/Yves/ImageSearchAi/Theme/default/components/molecules/image-uploader/README.md
@@ -1,0 +1,2 @@
+The component has been renamed to 'image-search-ai-image-uploader'.
+This change was necessary to address a naming collision that caused a compatibility issue with an existing Spryker module.
