--- src/SprykerEco/Yves/ImageSearchAi/Theme/default/components/molecules/image-uploader/index.ts
@@ -1,12 +1,0 @@
-import register from 'ShopUi/app/registry';
-import './image-uploader.scss';
-
-export default register(
-    'image-uploader',
-    () =>
-        import(
-            /* webpackMode: "lazy" */
-            /* webpackChunkName: "image-uploader" */
-            './image-uploader'
-            ),
-);
